Update source files in module
