+++ module/src/index.ts
@@ -15,14 +15,17 @@
 
 import { Request, RequestHandler, Response, Router } from 'express'
 import { Tracks as Tracks_, TrackAccumulator as TrackAccumulator_, TracksConfig } from './tracks'
-import { Context, LatLngTuple } from './types'
+import { Context, LatLngTuple, Position, VesselCollection, QueryParameters } from './types'
+import { validateParameters } from './utils'
 
 export interface ContextPosition {
   context: Context
-  value: {
-    latitude: number
-    longitude: number
-  }
+  value: Position
+}
+
+interface VesselTrack {
+  type: string,
+  coordinates: Array<Array<LatLngTuple>>
 }
 
 interface App {
@@ -38,6 +41,7 @@ interface App {
       onValue: (cb: (x: any) => void) => () => void
     }
   }
+  getSelfPath: (path: string) => void
 }
 
 interface Plugin {
@@ -55,6 +59,11 @@ interface Plugin {
 export default function ThePlugin(app: App): Plugin {
   let onStop: (() => void)[] = []
   let tracks: Tracks_ | undefined = undefined
+
+  function getVesselPosition():Position {
+    let p:any= app.getSelfPath('navigation.position');
+    return (p && p.value) ? p.value : null;
+  }
 
   return {
     start: function (configuration: TracksConfig) {
@@ -99,6 +108,31 @@ export default function ThePlugin(app: App): Plugin {
           })
       }
       router.get('/vessels/:vesselId/track', trackHandler.bind(this))
+
+      // return all / filtered vessel tracks
+      const allTracksHandler: RequestHandler = (req: Request, res: Response) => {
+        let params: QueryParameters= validateParameters(req.query)
+        app.debug('** params **', params)
+        tracks
+          ?.getAll(params, getVesselPosition())
+          .then((d: VesselCollection) => {
+            let trks:{[key: string] : VesselTrack}= {}
+            Object.entries(d).forEach( (i:[Context, LatLngTuple[]])=> {
+              trks[i[0]]= {
+                type: 'MultiLineString',
+                coordinates: [i[1]]
+              }
+            })
+            res.json(trks)
+          })
+          .catch(() => {
+            res.status(404)
+            res.json({ message: `No track available for vessels.`})
+          })
+      }
+      router.get('/tracks', allTracksHandler.bind(this) )
+      router.get('/tracks/*', allTracksHandler.bind(this) )
+
       return router
     },
 
@@ -125,8 +159,14 @@ export default function ThePlugin(app: App): Plugin {
           description: 'Tracks with no updates longer than this are removed',
           default: 600,
         },
-      },
-    },
+        maxRadius: {
+          type: 'number',
+          title: 'Maximum Radius (meters) ',
+          description: 'Include only vessels with position within this range. 0= all vessels',
+          default: 50000,
+        }
+      }
+    }
   }
 }
