--- src/types.ts
@@ -1,22 +1,0 @@
-export type Context = string
-
-export interface Position {
-  latitude: number
-  longitude: number
-}
-
-export interface Config {
-  resolution: number
-  pointsToKeep: number
-  maxAge: number,
-  maxRadius: number
-}
-
-export interface VesselCollection { [key: string] : Position[] }
-
-export interface GeoBounds {
-  ne: [number,number];
-  sw: [number,number];
-}
-
-export interface QueryParameters { [key: string] : any }
